% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_behavior_codes.R
\name{assign_behavior_codes}
\alias{assign_behavior_codes}
\title{Assign Behavior Codes to Signal Data}
\usage{
assign_behavior_codes(rec_dt, ana_dt)
}
\arguments{
\item{rec_dt}{A \code{data.table} with columns \code{time} and \code{signal}.}

\item{ana_dt}{A \code{data.table} with columns \code{behavior} and \code{start_time}.}
}
\value{
The \code{rec_dt} \code{data.table} with an added column \code{behavior}.
}
\description{
Assigns behavior codes to each time point in the signal data based on behavior intervals.
The \code{end_time} column is automatically created by shifting the \code{start_time} column,
assuming that the end time of each behavior is the start time of the next behavior.
}
\examples{
\dontrun{
rec_dt <- assign_behavior_codes(rec_dt, ana_dt)
}
}
